moved plots for tests to gallery Fragmment
