+++ android/api.yaml
@@ -1,0 +1,149 @@
+swagger: '2.0'
+info:
+  title: EatersLab API
+  version: v1
+host: eaterslab.herokuapp.com
+schemes:
+  - http
+basePath: /api
+consumes:
+  - application/json
+produces:
+  - application/json
+securityDefinitions:
+  Basic:
+    type: basic
+security:
+  - Basic: []
+paths:
+  /cafeteria/:
+    get:
+      operationId: cafeteria_list
+      description: ''
+      parameters: []
+      responses:
+        '200':
+          description: ''
+          schema:
+            type: array
+            items:
+              $ref: '#/definitions/Cafeteria'
+      tags:
+        - cafeteria
+    post:
+      operationId: cafeteria_create
+      description: ''
+      parameters:
+        - name: data
+          in: body
+          required: true
+          schema:
+            $ref: '#/definitions/Cafeteria'
+      responses:
+        '201':
+          description: ''
+          schema:
+            $ref: '#/definitions/Cafeteria'
+      tags:
+        - cafeteria
+    parameters: []
+  /cafeteria/{id}/:
+    get:
+      operationId: cafeteria_read
+      description: ''
+      parameters: []
+      responses:
+        '200':
+          description: ''
+          schema:
+            $ref: '#/definitions/Cafeteria'
+      tags:
+        - cafeteria
+    put:
+      operationId: cafeteria_update
+      description: ''
+      parameters:
+        - name: data
+          in: body
+          required: true
+          schema:
+            $ref: '#/definitions/Cafeteria'
+      responses:
+        '200':
+          description: ''
+          schema:
+            $ref: '#/definitions/Cafeteria'
+      tags:
+        - cafeteria
+    patch:
+      operationId: cafeteria_partial_update
+      description: ''
+      parameters:
+        - name: data
+          in: body
+          required: true
+          schema:
+            $ref: '#/definitions/Cafeteria'
+      responses:
+        '200':
+          description: ''
+          schema:
+            $ref: '#/definitions/Cafeteria'
+      tags:
+        - cafeteria
+    delete:
+      operationId: cafeteria_delete
+      description: ''
+      parameters: []
+      responses:
+        '204':
+          description: ''
+      tags:
+        - cafeteria
+    parameters:
+      - name: id
+        in: path
+        description: A unique integer value identifying this cafeteria.
+        required: true
+        type: integer
+definitions:
+  Cafeteria:
+    required:
+      - name
+      - description
+      - sub_description
+      - longitude
+      - latitude
+      - capacity
+    type: object
+    properties:
+      id:
+        title: ID
+        type: integer
+        readOnly: true
+      name:
+        title: Name
+        type: string
+        maxLength: 128
+        minLength: 1
+      description:
+        title: Description
+        type: string
+        maxLength: 256
+        minLength: 1
+      sub_description:
+        title: Sub description
+        type: string
+        maxLength: 512
+        minLength: 1
+      longitude:
+        title: Longitude
+        type: number
+      latitude:
+        title: Latitude
+        type: number
+      capacity:
+        title: Capacity
+        type: integer
+        maximum: 2147483647
+        minimum: -2147483648
