+++ android/app/src/main/java/labs/pooh/eaterslab/ui/gallery/GalleryFragment.kt
@@ -1,18 +1,22 @@
 package labs.pooh.eaterslab.ui.gallery
 
-import android.content.Context
-import android.content.Intent
-import android.graphics.Bitmap
 import android.os.Bundle
-import android.provider.MediaStore
+import android.view.Gravity
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
+import android.widget.FrameLayout
+import android.widget.Space
+import androidx.core.content.ContextCompat
+import androidx.core.view.plusAssign
 import androidx.fragment.app.Fragment
 import androidx.fragment.app.viewModels
-import androidx.lifecycle.Observer
 import kotlinx.android.synthetic.main.fragment_gallery.*
 import labs.pooh.eaterslab.R
+import labs.pooh.eaterslab.util.*
+import kotlin.math.abs
+import kotlin.math.log
+import kotlin.math.sin
 
 class GalleryFragment : Fragment() {
 
@@ -23,39 +27,47 @@ class GalleryFragment : Fragment() {
             container: ViewGroup?,
             savedInstanceState: Bundle?
     ): View? {
-
-        val root = inflater.inflate(R.layout.fragment_gallery, container, false)
-        galleryViewModel.text.observe(viewLifecycleOwner, Observer {
-            textGallery.text = it
-        })
-        return root
+        return inflater.inflate(R.layout.fragment_gallery, container, false)
     }
 
     override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
-        imageViewCamera.setOnClickListener { dispatchTakePictureIntent(view.context) }
         super.onViewCreated(view, savedInstanceState)
-    }
 
-    private fun dispatchTakePictureIntent(context: Context) {
-        Intent(MediaStore.ACTION_IMAGE_CAPTURE).also { intent ->
-            intent.resolveActivity(context.packageManager)?.also {
-                startActivityForResult(intent, REQUEST_IMAGE_CAPTURE)
-            }
-        }
-    }
+        val red = ContextCompat.getColor(requireContext(), R.color.colorAccent)
+        val yellow = ContextCompat.getColor(requireContext(), R.color.colorPrimaryDark)
 
-    override fun onActivityResult(requestCode: Int, resultCode: Int, data: Intent?) {
-        when(requestCode) {
-            REQUEST_IMAGE_CAPTURE -> {
-                (data?.extras?.get("data") as? Bitmap)?.let { imageBitmap ->
-                    imageViewCamera.setImageBitmap(imageBitmap)
+        val exampleData = listOf(
+            List(30) { it + sin(it.toDouble()) },
+            List(7) { 3 + sin(it.toDouble()) },
+            List(13) { abs(6 - it).toDouble() },
+            List(24) { if (it in (8..16)) (0..30).random().toDouble() else 0.0 },
+            List(13) { 12 - abs(6 - it).toDouble() },
+            List(10) { log(1.0 + it, 2.0) },
+            List(51) { sin(abs(it - 25).toDouble() + 0.001) / (abs(it - 25)  + 0.001) }
+        )
+
+        // example data plots
+        exampleData.forEach { dataSet ->
+            with(requireContext()) {
+                val frame = FrameLayout(this)
+                val barView = when(dataSet.size) {
+                    24 -> horizontalBarPlot(dataSet, red, ticks = getOrderedHours(), ticksIndexer = ::hoursIndexer, ticksScale = 2.0)
+                    30 -> horizontalBarPlot(dataSet, red, ticks = getOrderedMonthDays(4), ticksIndexer = ::monthDaysIndexer, ticksScale = 1.5)
+                    7 -> horizontalBarPlot(dataSet, red, ticks = getOrderedWeekDays(), ticksIndexer = this::weekDaysIndexer)
+                    else -> horizontalBarPlot(dataSet, red)
+                }
+                val plotView = discreteLinePlot(dataSet, yellow)
+                frame += barView
+                frame += plotView
+                frame.layoutParams = FrameLayout.LayoutParams(1000, 600).apply {
+                    gravity = Gravity.CENTER_HORIZONTAL
+                }
+
+                linearLayout += frame
+                linearLayout += Space(this).apply {
+                    layoutParams = ViewGroup.LayoutParams(0, 50)
                 }
             }
         }
-        super.onActivityResult(requestCode, resultCode, data)
-    }
-
-    companion object {
-        const val REQUEST_IMAGE_CAPTURE = 1
     }
 }
