--- android/app/src/main/res/layout/activity_plot.xml
@@ -1,27 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<androidx.constraintlayout.widget.ConstraintLayout xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:app="http://schemas.android.com/apk/res-auto"
-    xmlns:tools="http://schemas.android.com/tools"
-    android:layout_width="match_parent"
-    android:layout_height="match_parent"
-    tools:context=".PlotActivity">
-
-    <ScrollView
-        android:layout_width="match_parent"
-        android:layout_height="match_parent"
-        app:layout_constraintTop_toTopOf="parent"
-        app:layout_constraintStart_toEndOf="parent">
-
-        <LinearLayout
-            android:id="@+id/linearLayout"
-            android:layout_width="match_parent"
-            android:layout_height="wrap_content"
-            android:orientation="vertical"
-            android:gravity="center_horizontal"
-            android:paddingStart="@dimen/linear_layout_margin_padding"
-            android:paddingEnd="@dimen/linear_layout_margin_padding"
-            android:layout_marginTop="@dimen/linear_layout_margin_padding"
-            app:layout_constraintTop_toTopOf="parent"/>
-    </ScrollView>
-
-</androidx.constraintlayout.widget.ConstraintLayout>
